Implement pivot_table and crosstab functionality in C++ (#1688)
* Update to the testing framework.

* Correction.

* More implementations.

  Mathieu

* Removal of GetBit operations.

* Simplification of function calls.

* Introduction of the multiple_array_info data type.

* Removal of import pdb statements.

* More templatized code.

* Update.

* Introduce the sampling function f.

* More corrections for compilation.

* Resolution of the template problems.

* Some insert and debugging.

* More insertions.

* More bugs resolved.

* Current state of the code.

* Some bug resolved.

* Removal of cumulative / nunique / median
from the set of functions supported in pivot/crosstab.

* Another major code restructuration.

* Resolution of 3 problems with the compilation.

* Now it compiles.

* Correct one bug.

* More update to the code.

* Eliminate an unused code.

* Rename variable.

* Pass one test partially.

* Pass the first test.

* Major code removal.

* More code removal and insertion.

* One bug resolved with the removal of the columns by the compiler.

* Resolution of a painful bug.

* Debugging progress.

* The crosstab is working.

* Current state. Now needing to fix the index.

* Now address the problem of constructing the index.

* Put the set_columns_name_to_none and pass the test in serial.

* Removal of import pdb statements and simplification of the C++ code.

* One more bug solved. We pass the parallel test on one processor.

* Introduce the reorder_columns function.

* Change of the way index are accessed. First passing of the paralle
pivot_table.

* Restructuration of the cumulative string/list-string code.

* Restructuration of templatized code.

* Add documentation of options.

* More removal after merge.

* Resolve a template bug.

* Bugs resolved.

* Revert changes made to the templatized code.

* Yet another templatized var_agg statement put back.

* One more error after merge corrected.

* Make the system works for crosstab.

* Solve two memory leaks.

* Removal of unused code.

* Removal of import pdb statements.

* Some bugs in move resolved.

* Resolve more bugs in crosstab.

* Correction on the tests.

* Support of min/max/sum/prod for integers. Bugs solved.

* Introduction of mean/var/std functionality for the pivot_table.

* Restructuration of the tests.

* Correct the test.

* Resolve some memory leaks.

* More cleaning of debugging stuff.

* Apply blakc and clang-format.

* Removal of debug statements.

* Some correction for compilation.

* Add some docs.

* Resolve issues pointed out.

* Update bodo/tests/utils.py


* Correction after compilation.

* Removal of one merge leftover.

+++ bodo/libs/_shuffle.cpp
@@ -1024,8 +1024,8 @@ array_info* reverse_shuffle_numpy_array(array_info* in_arr, uint32_t* hashes,
                                         mpi_comm_info const& comm_info) {
     uint64_t siztype = numpy_item_size[in_arr->dtype];
     MPI_Datatype mpi_typ = get_MPI_typ(in_arr->dtype);
-    int64_t n_rows_ret = std::accumulate(comm_info.send_count.begin(),
-                                         comm_info.send_count.end(), 0);
+    size_t n_rows_ret = std::accumulate(comm_info.send_count.begin(),
+                                        comm_info.send_count.end(), 0);
     array_info* out_arr = alloc_array(n_rows_ret, 0, 0, in_arr->arr_type,
                                       in_arr->dtype, 0, in_arr->num_categories);
     char* data1_i = in_arr->data1;
